+++ CHANGELOG.md
@@ -13,6 +13,7 @@ For a full diff see [`1.0.0...main`][1.0.0...main].
 - Dropped support for PHP 7.3 ([#137]), by [@localheinz]
 - Renamed `Json::encoded()` to `Json::toString()` ([#155]), by [@localheinz]
 - Inlined `Decoder` into `SchemaValidator` ([#157]), by [@localheinz]
+- Disallowed injection of `Validator` into `SchemaValidator` ([#158]), by [@localheinz]
 
 ## [`1.0.0`][1.0.0]
 
@@ -39,5 +40,6 @@ For a full diff see [`dcd4cfb...1.0.0`][dcd4cfb...1.0.0].
 [#137]: https://github.com/ergebnis/json-schema-validator/pull/137
 [#155]: https://github.com/ergebnis/json-schema-validator/pull/155
 [#157]: https://github.com/ergebnis/json-schema-validator/pull/157
+[#158]: https://github.com/ergebnis/json-schema-validator/pull/158
 
 [@localheinz]: https://github.com/localheinz